tools/requirements.txt
pyserial
hidapi
